moving all actions to the database level.
This is a branch point where we move all action to the database level.
By actions we mean

   (1) Display of entities either inline or in selection list
   (2) Access control etc.

If these actions were at the Handler level, then stuff will happen in
different transactions. Consider the example of deleting a list of
objects given. For each object in a selection, we should check whether
the given user has permission to delete the given object or not. If
this happened in the Handler level, we would have to first pull out
all objects via a runDB, test whether the user has permissions, then
run the dbaction to delete the object. So this will not be a single
tranction and hence both inefficient and possibly incorrect.

We now move all these details to the DB level. There is some loss of
generality in the choice.