merging the crud and selection subsites
Clearly if the persist backend of a master site is not an instance of
PersistQuery there is now abstract way of supporting the selection
pages. This is why we split the Crud and Selection subsite. However,
this meant that generating links to the crud pages becomes
complicated. So we have merged the definitions again. We will take care
of this complication at time of defining the Dispatch instance.