split the crud and selection subsites again
The problem with a single admin subsite was that depending on whether
the backend supports query or not the YesodDispatch instance have to
change. Although there is a way of handling this via the language
extension OverlappingInstances and functional dependencies, this was
not desirable because:

1. It involves deep magic:
    www.haskell.org/haskellwiki/GHC/AdvancedOverlap

2. There are dummy routes like ListR should give some default error
   pages that says that the page does not exist. Why then have those
   routes. It in some sense defeats the purpose of type safe urls: A
   user might give use ListR in her template which when clicked gives
   an error page and this will not be detected at compile time.

The hackish solution that we implement is to have an additional class
called LiftCrudRoutes which can be defined only when the admin sites
are hooked. If we use standard TH functions, we can generate an
instance for this.