preparing for i18n
In admin sites there are few things that depend on i18n. They are

   1. The messages like delete, modify etc and
   2. Entity specific stuff like Attribute titles and Action names.

Clearly translating contents of attributes is beyound the scope.

The yesod way to handle this is via the class RenderMessage. We had an
interfaces where stuff like attribute titles were hard coded into the
class Administrable. We now take the approach of declaring an instance
of RenderMessage for both Attribute v and Action v. This solves two
problems in one shot.

1. Overriding the default titles/action names.
2. Proper i18n.

This also considerably simplified the TH code for entity.