predefined Admin subsite that hooks all crud/selection subsites
The strategy used by the TH code to hook all our crud and selection
sites was by creating a new subsite, say SiteAdmin (when Site is our
foundation type). The user can then hook this subtype SiteAdmin into
the main routes (i.e. the routes of Site). This has the following
problem: Although the TH code knows how to lift the crud and selection
routes to SiteAdmin, as all the corresponding type safe url
constructors are generated, it does not know how to lift the admin
site routes. This is the reason why function like mkAdminSite had an
extra argument (the string representation of the type safe url). In
this branch, we follow an alternate strategy. We predefine a type
Admin master which will be the final admin subsite onto which the crud
and selection subsites are hooked. The user of the library will have
to then define the LiftAdminRoutes class which will lift all admin
routes. Our template haskell functions now need not take the extra
argument.

There is a trade-off here: user need to define one more type class and
thus more complicated usage. However, the advantage in the approach is
better error message: A typos in the extra string argument will be
caught only when compiling the generated code and thus can give very
complicated errors.